Module rstream::channel

Wrapper to in-memory channels.

This module exists to ease the transition between channel libraries.

Structs

BoundedChannelReceiver
A wrapper on a bounded channel receiver.

BoundedChannelSender
A wrapper on a bounded channel sender.

SelectAnyResult
The result of a select_any.

UnboundedChannelReceiver
A wrapper on an unbounded channel receiver.

UnboundedChannelSender
A wrapper on an unbounded channel sender.

Enums

SelectResult
An either type with the result of a select on 2 channels.

Traits

ChannelItem

Type Definitions

RecvError

RecvTimeoutError

TryRecvError