Module rstream::operator

Operators that can be applied to a stream.

The actual operator list can be found from the implemented methods of Stream, KeyedStream, WindowedStream and KeyedWindowedStream.

Modules

join

Some utility structures for building the join operators.

sink

Utility traits and structures related to the sink operators.

source

Utility traits and structures related to the source operators.

window

The types related to the windowed streams.

Enums

StreamElement

An element of the stream. This is what enters and exits from the operators.

Traits

Data

Marker trait that all the types inside a stream should implement.

DataKey

Marker trait that all the keys should implement.

ExchangeData

Marker trait for data types that are used to communicate between different blocks.

ExchangeDataKey

Marker trait for key types that are used when communicating between different blocks.

KeyerFn

Marker trait for the function that extracts the key out of a type.

Operator

An operator represents a unit of computation. It’s always included inside a chain of operators, inside a block.

Type Definitions

Timestamp

When using timestamps and watermarks, this type expresses the timestamp of a message or of a watermark.