Struct rstream::operator::join::JoinStreamLocalSortMerge

pub struct JoinStreamLocalSortMerge<Key: Data + Ord, Out1: ExchangeData, Out2: ExchangeData, Keyer1: KeyerFn<Key, Out1>, Keyer2: KeyerFn<Key, Out2>, ShipStrat: ShipStrategy> { /* fields omitted */ }

This is an intermediate type for building a join operator.

The ship strategy has already been selected and it’s stored in ShipStrat, the local strategy is hash and now the join variant has to be selected.

Note that outer join is not supported if the ship strategy is broadcast_right.

Implementations

impl<Key: Data + Ord, Out1: ExchangeData, Out2: ExchangeData, Keyer1, Keyer2> JoinStreamLocalSortMerge<Key, Out1, Out2, Keyer1, Keyer2, ShipHash> where
Keyer1: KeyerFn<Key, Out1>,
Keyer2: KeyerFn<Key, Out2>,

pub fn inner(
self
) -> KeyedStream<Key, InnerJoinTuple<Out1, Out2>, impl Operator<KeyValue<Key, InnerJoinTuple<Out1, Out2>>>>

Finalize the join operator by specifying that this is an inner join.

Given two stream, create a stream with all the pairs (left item from the left stream, right item from the right), such that the key obtained with keyer1 on an item from the left is equal to the key obtained with keyer2 on an item from the right.

This is an inner join, very similarly to SELECT a, b FROM a JOIN b ON keyer1(a) = keyer2(b).

Note: this operator will split the current block.

pub fn left(
self
) -> KeyedStream<Key, LeftJoinTuple<Out1, Out2>, impl Operator<KeyValue<Key, LeftJoinTuple<Out1, Out2>>>>

Finalize the join operator by specifying that this is a left join.

Given two stream, create a stream with all the pairs (left item from the left stream, right item from the right), such that the key obtained with keyer1 on an item from the left is equal to the key obtained with keyer2 on an item from the right.

This is a left join, meaning that if an item from the left does not find and element from the right with which make a pair, an extra pair (left, None) is generated. If you want to have a right join, you just need to switch the two sides and use a left join.

This is very similar to SELECT a, b FROM a LEFT JOIN b ON keyer1(a) = keyer2(b).

Note: this operator will split the current block.

pub fn outer(
self
) -> KeyedStream<Key, OuterJoinTuple<Out1, Out2>, impl Operator<KeyValue<Key, OuterJoinTuple<Out1, Out2>>>>

Finalize the join operator by specifying that this is an outer join.

Given two stream, create a stream with all the pairs (left item from the left stream, right item from the right), such that the key obtained with keyer1 on an item from the left is equal to the key obtained with keyer2 on an item from the right.

This is a full-outer join, meaning that if an item from the left does not find and element from the right with which make a pair, an extra pair (left, None) is generated. Similarly if an element from the right does not appear in any pair, a new one is generated with (None, right).

This is very similar to SELECT a, b FROM a FULL OUTER JOIN b ON keyer1(a) = keyer2(b).

Note: this operator will split the current block.

impl<Key: Data + Ord, Out1: ExchangeData, Out2: ExchangeData, Keyer1, Keyer2> JoinStreamLocalSortMerge<Key, Out1, Out2, Keyer1, Keyer2, ShipBroadcastRight> where
Keyer1: KeyerFn<Key, Out1>,
Keyer2: KeyerFn<Key, Out2>,

pub fn inner(
self
) -> Stream<KeyValue<Key, InnerJoinTuple<Out1, Out2>>, impl Operator<KeyValue<Key, InnerJoinTuple<Out1, Out2>>>>

Finalize the join operator by specifying that this is an inner join.

Given two stream, create a stream with all the pairs (left item from the left stream, right item from the right), such that the key obtained with keyer1 on an item from the left is equal to the key obtained with keyer2 on an item from the right.

This is an inner join, very similarly to SELECT a, b FROM a JOIN b ON keyer1(a) = keyer2(b).

Note: this operator will split the current block.

pub fn left(
self
) -> Stream<KeyValue<Key, LeftJoinTuple<Out1, Out2>>, impl Operator<KeyValue<Key, LeftJoinTuple<Out1, Out2>>>>

Finalize the join operator by specifying that this is a left join.

Given two stream, create a stream with all the pairs (left item from the left stream, right item from the right), such that the key obtained with keyer1 on an item from the left is equal to the key obtained with keyer2 on an item from the right.

This is a left join, meaning that if an item from the left does not find and element from the right with which make a pair, an extra pair (left, None) is generated. If you want to have a right join, you just need to switch the two sides and use a left join.

This is very similar to SELECT a, b FROM a LEFT JOIN b ON keyer1(a) = keyer2(b).

Note: this operator will split the current block.