Trait rstream::operator::WindowGenerator

pub trait WindowGenerator<Key: DataKey, Out: Data>: Send {
    fn add(&mut self, item: StreamElement<KeyValue<Key, Out>>);

    fn next_window(&mut self) -> Option<Window<'_, Key, Out>>;

    fn advance(&mut self);

    fn buffer(&self) -> &VecDeque<Out>;
}

A WindowGenerator handles the generation of windows for a given key.

Required methods

fn add(&mut self, item: StreamElement<KeyValue<Key, Out>>)

Handle a new element of the stream.

fn next_window(&mut self) -> Option<Window<'_, Key, Out>>

If a window is ready, return it so that it can be processed.

fn advance(&mut self)

Close the current open window. This method is called when a Window is dropped after being processed.

fn buffer(&self) -> &VecDeque<Out>

Return the buffer from which Window will get the elements of the window.

Loading content...