Module rstream::operator::window

The types related to the windowed streams.

Structs

CountWindow
Count windows divide elements into windows by counting them. They can be used with both timestamped and non-timestamped streams. When used with timestamped streams, elements are ordered according to their timestamps.

EventTimeWindow
Event-time windows divide elements into windows based on their timestamp. For this reason, event-time windows only support timestamped elements and cannot be used on streams of elements without a timestamp.

ProcessingTimeWindow
Processing-time windows divide elements into windows based on their arrival time. The elements’ timestamps are assigned automatically using the system time of the machine. For this reason, processing-time windows only support non-timestamped elements and cannot be used on streams of elements with timestamps and watermarks.

SessionEventTimeWindowDescr
Window description for session event time windows

SessionProcessingTimeWindowDescr
Window description for session processing time windows

SlidingEventTimeWindowDescr
Window description for sliding event time windows

SlidingProcessingTimeWindowDescr
Window description for sliding processing time windows

Window
A window is a collection of elements and may be associated with a timestamp.

Traits

WindowDescription
A WindowDescription describes how a window behaves.

WindowGenerator
A WindowGenerator handles the generation of windows for a given key.