Trait rstream::operator::window::WindowDescription

pub trait WindowDescription<Key: DataKey, Out: Data>: Send {
    type Generator: WindowGenerator<Key, Out> + Clone + 'static;
    fn new_generator(&self) -> Self::Generator;

    fn to_string(&self) -> String;
}

A WindowDescription describes how a window behaves.

Associated Types

type Generator: WindowGenerator<Key, Out> + Clone + 'static

The type of the window generator of this window.

Loading content...
Required methods

fn new_generator(&self) -> Self::Generator

Construct a new window generator, ready to handle elements.

fn to_string(&self) -> String

Loading content...