Struct rstream::stream::KeyedStream

pub struct KeyedStream<Key: DataKey, Out: Data, OperatorChain>(pub Stream<KeyValue<Key, Out>, OperatorChain>)
 where
    OperatorChain: Operator<KeyValue<Key, Out>>;

A KeyedStream is like a set of Streams, each of which partitioned by some Key. Internally it's just a stream whose elements are KeyValue pairs and the operators behave following the KeyedStream semantics.

The type of the Key must be a valid key inside an hashmap.

Implementations

impl<Key: DataKey, Out: Data, OperatorChain> KeyedStream<Key, Out, OperatorChain> where
OperatorChain: Operator<KeyValue<Key, Out>> + 'static,

pub fn batch_mode(self, batch_mode: BatchMode) -> Self

impl<Key: DataKey, Out: Data, OperatorChain> KeyedStream<Key, Out, OperatorChain> where
OperatorChain: Operator<KeyValue<Key, Out>> + 'static,

pub fn concat<OperatorChain2>(
self,
oth: KeyedStream<Key, Out, OperatorChain2>
) -> KeyedStream<Key, Out, impl Operator<KeyValue<Key, Out>>> where
OperatorChain2: Operator<KeyValue<Key, Out>> + 'static,

impl<Key: DataKey, Out: Data, OperatorChain> KeyedStream<Key, Out, OperatorChain> where
OperatorChain: Operator<KeyValue<Key, Out>> + 'static,

pub fn filter<Predicate>(
self,
predicate: Predicate
) -> KeyedStream<Key, Out, impl Operator<KeyValue<Key, Out>>> where
Predicate: Fn(&KeyValue<Key, Out>) -> bool + Clone + Send + 'static,

impl<Key: DataKey, Out: Data, OperatorChain> KeyedStream<Key, Out, OperatorChain> where
OperatorChain: Operator<KeyValue<Key, Out>> + 'static,

pub fn filter_map<NewOut: Data, F>(
self,
f: F
) -> KeyedStream<Key, NewOut, impl Operator<KeyValue<Key, NewOut>>> where
F: Fn(KeyValue<&Key, Out>) -> Option<NewOut> + Send + Clone + 'static,

impl<Key: DataKey, Out: Data, NewOut: Data, OperatorChain> KeyedStream<Key, Out, OperatorChain> where
Out: IntoIterator<Item = NewOut>,
<Out as IntoIterator>::IntoIter: Clone + Send + 'static,
OperatorChain: Operator<KeyValue<Key, Out>> + 'static,

pub fn flatten(
self
) -> KeyedStream<Key, NewOut, impl Operator<KeyValue<Key, NewOut>>>

impl<Key: DataKey, Out: Data, OperatorChain> KeyedStream<Key, Out, OperatorChain> where
OperatorChain: Operator<KeyValue<Key, Out>> + 'static,

pub fn flat_map<NewOut: Data, MapOut: Data, F>(
self,
f: F
) -> KeyedStream<Key, NewOut, impl Operator<KeyValue<Key, NewOut>>> where
MapOut: IntoIterator<Item = NewOut>,
<MapOut as IntoIterator>::IntoIter: Clone + Send + 'static,
F: Fn(KeyValue<&Key, Out>) -> MapOut + Send + Clone + 'static,

impl<Key: DataKey, Out: Data, OperatorChain> KeyedStream<Key, Out, OperatorChain> where
OperatorChain: Operator<KeyValue<Key, Out>> + 'static,

pub fn fold<NewOut: Data, F>(
self,
init: NewOut,
f: F
) -> KeyedStream<Key, NewOut, impl Operator<KeyValue<Key, NewOut>>> where
F: Fn(&mut NewOut, Out) + Send + Clone + 'static,

impl<Key: DataKey, Out: Data, OperatorChain> KeyedStream<Key, Out, OperatorChain> where
OperatorChain: Operator<KeyValue<Key, Out>> + 'static,

pub fn reduce<F>(
self,
f: F
) -> KeyedStream<Key, Out, impl Operator<KeyValue<Key, Out>>> where
F: Fn(&mut Out, Out) + Send + Clone + 'static,

impl<Key: DataKey, Out: Data, OperatorChain> KeyedStream<Key, Out, OperatorChain> where
OperatorChain: Operator<KeyValue<Key, Out>> + 'static,

pub fn map<NewOut: Data, F>(
self,
f: F
) -> KeyedStream<Key, NewOut, impl Operator<KeyValue<Key, NewOut>>> where
F: Fn(KeyValue<&Key, Out>) -> NewOut + Send + Clone + 'static,

impl<Key: DataKey, Out: Data, OperatorChain> KeyedStream<Key, Out, OperatorChain> where
OperatorChain: Operator<KeyValue<Key, Out>> + 'static,

pub fn for_each<F>(self, f: F) where
F: Fn(Key, Out) + Send + Clone + 'static,

impl<Key: DataKey, Out: Data, OperatorChain> KeyedStream<Key, Out, OperatorChain> where
OperatorChain: Operator<KeyValue<Key, Out>> + 'static,

pub fn unkey(
self
) -> Stream<KeyValue<Key, Out>, impl Operator<KeyValue<Key, Out>>>

impl<Key: DataKey, Out: Data, OperatorChain> KeyedStream<Key, Out, OperatorChain> where
OperatorChain: Operator<KeyValue<Key, Out>> + 'static,

pub fn window<WinOut: Data, WinDescr: WindowDescription<Key, WinOut>>(
self,
descr: WinDescr
) -> KeyedWindowedStream<Key, Out, impl Operator<KeyValue<Key, Out>>, WinOut, WinDescr>