Struct rstream::stream::Stream

pub struct Stream<Out: Data, OperatorChain> where
    OperatorChain: Operator<Out>,  { /* fields omitted */ }

A Stream represents a chain of operators that work on a flow of data. The type of the elements entering the stream is In, while the type of the outgoing elements is Out.

Internally a stream is composed by a chain of blocks, each of which can be seen as a simpler stream with input and output types.

A block is internally composed of a chain of operators, nested like the Iterator from std. The type of the chain inside the block is OperatorChain and it's required as type argument of the stream. This type only represents the chain inside the last block of the stream, not all the blocks inside of it.

Implementations

impl<Out: Data, OperatorChain> Stream<Out, OperatorChain> where
OperatorChain: Operator<Out> + Send + 'static,

pub fn batch_mode(self, batch_mode: BatchMode) -> Self

impl<Out: Data, OperatorChain> Stream<Out, OperatorChain> where
OperatorChain: Operator<Out> + Send + 'static,

pub fn concat<OperatorChain2>(
self,
oth: Stream<Out, OperatorChain2>
) -> Stream<Out, impl Operator<Out>> where
OperatorChain2: Operator<Out> + Send + 'static,

impl<Out: Data, OperatorChain> Stream<Out, OperatorChain> where
OperatorChain: Operator<Out> + Send + 'static,

pub fn filter<Predicate>(
self,
predicate: Predicate
) -> Stream<Out, impl Operator<Out>> where
Predicate: Fn(&Out) -> bool + Clone + 'static,

impl<Out: Data, OperatorChain> Stream<Out, OperatorChain> where
OperatorChain: Operator<Out> + Send + 'static,

pub fn filter_map<NewOut: Data, F>(
self,
f: F
) -> Stream<NewOut, impl Operator<NewOut>> where
F: Fn(Out) -> Option<NewOut> + Send + Sync + 'static,

impl<Out: Data, OperatorChain, NewOut: Data> Stream<Out, OperatorChain> where
Out: IntoIterator<Item = NewOut>,
<Out as IntoIterator>::IntoIter: Clone + Send + 'static,
OperatorChain: Operator<Out> + Send + 'static,

pub fn flatten(self) -> Stream<NewOut, impl Operator<NewOut>>

impl<Out: Data, OperatorChain> Stream<Out, OperatorChain> where
OperatorChain: Operator<Out> + Send + 'static,

pub fn flat_map<MapOut: Data, NewOut: Data, F>(
self,
f: F
) -> Stream<NewOut, impl Operator<NewOut>> where
MapOut: IntoIterator<Item = NewOut>,
<MapOut as IntoIterator>::IntoIter: Clone + Send + 'static,
F: Fn(Out) -> MapOut + Send + Sync + 'static,

impl<Out: Data, OperatorChain> Stream<Out, OperatorChain> where
OperatorChain: Operator<Out> + Send + 'static,

pub fn fold<NewOut: Data, F>(
self,
init: NewOut,
f: F
) -> Stream<NewOut, impl Operator<NewOut>> where
F: Fn(&mut NewOut, Out) + Send + Sync + 'static,

pub fn fold_assoc<NewOut: Data, Local, Global>(
self,
init: NewOut,
local: Local,
global: Global
) -> Stream<NewOut, impl Operator<NewOut>> where
Local: Fn(&mut NewOut, Out) + Send + Sync + 'static,
Global: Fn(&mut NewOut, NewOut) + Send + Sync + 'static,

impl<Out: Data, OperatorChain> Stream<Out, OperatorChain> where
OperatorChain: Operator<Out> + Send + 'static,

pub fn group_by<Key: DataKey, Keyer>(
self,
keyer: Keyer
) -> KeyedStream<Key, Out, impl Operator<KeyValue<Key, Out>>> where
Keyer: Fn(&Out) -> Key + Send + Sync + 'static,

impl<Out: Data, OperatorChain> Stream<Out, OperatorChain> where
OperatorChain: Operator<Out> + Send + 'static,

pub fn iterate<Body, DeltaUpdate: Data + Default, State: Data, OperatorChain2>(
self,
num_iterations: usize,
initial_state: State,
body: Body,
local_fold: impl Fn(&mut DeltaUpdate, Out) + Send + Sync + 'static,
global_fold: impl Fn(&mut State, DeltaUpdate) + Send + Sync + 'static,
loop_condition: impl Fn(&mut State) -> bool + Send + Sync + 'static
) -> (Stream<State, impl Operator<State>>, Stream<Out, impl Operator<Out>>) where
Body: Fn(Stream<Out, Iterate<Out, State, OperatorChain>>, IterationStateHandle<State>) -> Stream<Out, OperatorChain2>,
OperatorChain2: Operator<Out> + Send + 'static,

impl<Out: Data, OperatorChain> Stream<Out, OperatorChain> where
OperatorChain: Operator<Out> + Send + 'static,

pub fn replay<Body, DeltaUpdate: Data + Default, State: Data, OperatorChain2>(
self,
num_iterations: usize,
initial_state: State,
body: Body,
local_fold: impl Fn(&mut DeltaUpdate, Out) + Send + Sync + 'static,
global_fold: impl Fn(&mut State, DeltaUpdate) + Send + Sync + 'static,
loop_condition: impl Fn(&mut State) -> bool + Send + Sync + 'static
) -> Stream<State, impl Operator<State>> where
Body: Fn(Stream<Out, Replay<Out, State, OperatorChain>>, IterationStateHandle<State>) -> Stream<Out, OperatorChain2>,
OperatorChain2: Operator<Out> + Send + 'static,

impl<Out: Data, OperatorChain> Stream<Out, OperatorChain> where
OperatorChain: Operator<Out> + Send + 'static,

pub fn key_by<Key: DataKey, Keyer>(
self,
keyer: Keyer
) -> KeyedStream<Key, Out, impl Operator<KeyValue<Key, Out>>> where
Keyer: Fn(&Out) -> Key + Send + Sync + 'static,

impl<Out: Data, OperatorChain> Stream<Out, OperatorChain> where
OperatorChain: Operator<Out> + Send + 'static,

pub fn group_by_fold<Key: DataKey, NewOut: Data, Keyer, Local, Global>(
self,
keyer: Keyer,
init: NewOut,
local: Local,
global: Global
) -> KeyedStream<Key, NewOut, impl Operator<KeyValue<Key, NewOut>>> where
Keyer: Fn(&Out) -> Key + Send + Sync + 'static,
Local: Fn(&mut NewOut, Out) + Send + Sync + 'static,
Global: Fn(&mut NewOut, NewOut) + Send + Sync + 'static,

impl<Out: Data, OperatorChain> Stream<Out, OperatorChain> where
OperatorChain: Operator<Out> + Send + 'static,

pub fn group_by_reduce<Key: DataKey, Keyer, F>(
self,
keyer: Keyer,
f: F
) -> KeyedStream<Key, Out, impl Operator<KeyValue<Key, Out>>> where
Keyer: Fn(&Out) -> Key + Send + Sync + 'static,
F: Fn(&mut Out, Out) + Send + Sync + 'static,

impl<Out: Data, OperatorChain> Stream<Out, OperatorChain> where
OperatorChain: Operator<Out> + Send + 'static,

pub fn map<NewOut: Data, F>(self, f: F) -> Stream<NewOut, impl Operator<NewOut>> where
F: Fn(Out) -> NewOut + Send + Sync + 'static,

impl<Out: Data, OperatorChain> Stream<Out, OperatorChain> where
OperatorChain: Operator<Out> + Send + 'static,

pub fn max_parallelism(
self,
max_parallelism: usize
) -> Stream<Out, impl Operator<Out>>

impl<Out: Data, OperatorChain> Stream<Out, OperatorChain> where
OperatorChain: Operator<Out> + Send + 'static,

pub fn reduce<F>(self, f: F) -> Stream<Out, impl Operator<Out>> where
F: Fn(&mut Out, Out) + Send + Sync + 'static,

pub fn reduce_assoc<F>(self, f: F) -> Stream<Out, impl Operator<Out>> where
F: Fn(&mut Out, Out) + Send + Sync + 'static,

impl<Out: Data, OperatorChain> Stream<Out, OperatorChain> where
OperatorChain: Operator<Out> + Send + 'static,

pub fn shuffle(self) -> Stream<Out, impl Operator<Out>>

impl<Out: Data, OperatorChain> Stream<Out, OperatorChain> where
OperatorChain: Operator<Out> + Send + 'static,

pub fn collect_vec(self) -> StreamOutput<Vec<Out>>

impl<Out: Data, OperatorChain> Stream<Out, OperatorChain> where
OperatorChain: Operator<Out> + Send + 'static,

pub fn for_each<F: Fn(Out) + Send + Sync + 'static>(self, f: F)

impl<Out: Data, OperatorChain> Stream<Out, OperatorChain> where
OperatorChain: Operator<Out> + Send + 'static,

pub fn split(self, splits: usize) -> Vec<Stream<Out, impl Operator<Out>>>

impl<Out: Data, OperatorChain> Stream<Out, OperatorChain> where
OperatorChain: Operator<Out> + Send + 'static,

pub fn window_all<WinOut: Data, WinDescr: WindowDescription<(), WinOut>>(
self,
descr: WinDescr
) -> WindowedStream<Out, impl Operator<KeyValue<(), Out>>, WinOut, WinDescr>

impl<Out1: Data, OperatorChain1> Stream<Out1, OperatorChain1> where
OperatorChain1: Operator<Out1> + Send + 'static,

pub fn zip<Out2: Data, OperatorChain2>(
self,
oth: Stream<Out2, OperatorChain2>
) -> Stream<(Out1, Out2), impl Operator<(Out1, Out2)>> where
OperatorChain2: Operator<Out2> + Send + 'static,