Struct rstream::KeyedStream

pub struct KeyedStream<Key: Data, Out: Data, OperatorChain>(pub Stream<KeyValue<Key, Out>, OperatorChain>)
 where
    OperatorChain: Operator<KeyValue<Key, Out>>;

A KeyedStream is like a set of Streams, each of which partitioned by some Key. Internally it’s just a stream whose elements are KeyValue pairs and the operators behave following the KeyedStream semantics.

The type of the Key must be a valid key inside an hashmap.

Implementations

impl<Key: DataKey, Out: Data, OperatorChain> KeyedStream<Key, Out, OperatorChain> where
OperatorChain: Operator<KeyValue<Key, Out>> + 'static,

pub fn batch_mode(self, batch_mode: BatchMode) -> Self

Change the batch mode for this stream.

This change will be propagated to all the operators following, even of the next blocks, until it’s changed again.

Example

use rstream::BatchMode;

let s = env.stream(IteratorSource::new((0..10))).group_by(|&n| n % 2);
s.batch_mode(BatchMode::fixed(1024));

impl<Key: ExchangeDataKey, Out: ExchangeData, OperatorChain> KeyedStream<Key, Out, OperatorChain> where
OperatorChain: Operator<KeyValue<Key, Out>> + 'static,

pub fn concat<OperatorChain2>(
self,
oth: KeyedStream<Key, Out, OperatorChain2>
) -> KeyedStream<Key, Out, impl Operator<KeyValue<Key, Out>>> where
OperatorChain2: Operator<KeyValue<Key, Out>> + 'static,

Concatenate the items of this stream with the items of another stream with the same type.

Note: the order of the resulting items is not specified.

Note: this operator will split the current block.

Example

let s1 = env.stream(IteratorSource::new((0..3))).group_by(|&n| n % 2);
let s2 = env.stream(IteratorSource::new((3..5))).group_by(|&n| n % 2);
let res = s1.concat(s2).collect_vec();

env.execute();

let mut res = res.get().unwrap();
res.sort_unstable(); // the output order is nondeterministic
assert_eq!(res, vec![(0, 0), (0, 2), (0, 4), (1, 1), (1, 3)]);

impl<Key: DataKey, Out: Data, OperatorChain> KeyedStream<Key, Out, OperatorChain> where
OperatorChain: Operator<KeyValue<Key, Out>> + 'static,

pub fn filter<Predicate>(
self,
predicate: Predicate
) -> KeyedStream<Key, Out, impl Operator<KeyValue<Key, Out>>> where
Predicate: Fn(&KeyValue<Key, Out>) -> bool + Clone + Send + 'static,

Remove from the stream all the elements for which the provided predicate returns false.

Note: this is very similar to Iteartor::filter

Example

let s = env.stream(IteratorSource::new((0..10))).group_by(|&n| n % 2);
let res = s.filter(|&(_key, n)| n % 3 == 0).collect_vec();

env.execute();

let mut res = res.get().unwrap();
res.sort_unstable();
assert_eq!(res, vec![(0, 0), (0, 6), (1, 3), (1, 9)]);

impl<Key: DataKey, Out: Data, OperatorChain> KeyedStream<Key, Out, OperatorChain> where
OperatorChain: Operator<KeyValue<Key, Out>> + 'static,

pub fn filter_map<NewOut: Data, F>(
self,
f: F
) -> KeyedStream<Key, NewOut, impl Operator<KeyValue<Key, NewOut>>> where
F: Fn(KeyValue<&Key, Out>) -> Option<NewOut> + Send + Clone + 'static,

Remove from the stream all the elements for which the provided function returns None and keep the elements that returned Some(_).

Note: this is very similar to Iteartor::filter_map

Example

let s = env.stream(IteratorSource::new((0..10))).group_by(|&n| n % 2);
let res = s.filter_map(|(_key, n)| if n % 3 == 0 { Some(n * 4) } else { None }).collect_vec();

env.execute();

let mut res = res.get().unwrap();
res.sort_unstable();
assert_eq!(res, vec![(0, 0), (0, 24), (1, 12), (1, 36)]);

impl<Key: DataKey, Out: Data, NewOut: Data, OperatorChain> KeyedStream<Key, Out, OperatorChain> where
Out: IntoIterator<Item = NewOut>,
<Out as IntoIterator>::IntoIter: Clone + Send + 'static,
OperatorChain: Operator<KeyValue<Key, Out>> + 'static,

pub fn flatten(
self
) -> KeyedStream<Key, NewOut, impl Operator<KeyValue<Key, NewOut>>>

Transform this stream of containers into a stream of all the contained values.

Note: this is very similar to Iteartor::flatten

Example

let s = env
    .stream(IteratorSource::new(vec![
        vec![0, 1, 2],
        vec![3, 4, 5],
        vec![6, 7]
    ].into_iter()))
    .group_by(|v| v[0] % 2);
let res = s.flatten().collect_vec();

env.execute();

let mut res = res.get().unwrap();
res.sort_unstable();
assert_eq!(res, vec![(0, 0), (0, 1), (0, 2), (0, 6), (0, 7), (1, 3), (1, 4), (1, 5)]);

impl<Key: DataKey, Out: Data, OperatorChain> KeyedStream<Key, Out, OperatorChain> where
OperatorChain: Operator<KeyValue<Key, Out>> + 'static,

pub fn flat_map<NewOut: Data, MapOut: Data, F>(
self,
f: F
) -> KeyedStream<Key, NewOut, impl Operator<KeyValue<Key, NewOut>>> where
MapOut: IntoIterator<Item = NewOut>,
<MapOut as IntoIterator>::IntoIter: Clone + Send + 'static,
F: Fn(KeyValue<&Key, Out>) -> MapOut + Send + Clone + 'static,

Apply a mapping operation to each element of the stream, the resulting stream will be the flattened values of the result of the mapping.

Note: this is very similar to Iteartor::flat_map.

Example

let s = env.stream(IteratorSource::new((0..3))).group_by(|&n| n % 2);
let res = s.flat_map(|(_key, n)| vec![n, n]).collect_vec();

env.execute();

let mut res = res.get().unwrap();
res.sort_unstable();
assert_eq!(res, vec![(0, 0), (0, 0), (0, 2), (0, 2), (1, 1), (1, 1)]);

impl<Key: ExchangeDataKey, Out: ExchangeData, OperatorChain> KeyedStream<Key, Out, OperatorChain> where
OperatorChain: Operator<KeyValue<Key, Out>> + 'static,

pub fn interval_join<Out2, OperatorChain2>(
self,
right: KeyedStream<Key, Out2, OperatorChain2>,
lower_bound: Duration,
upper_bound: Duration
) -> KeyedStream<Key, (Out, Out2), impl Operator<KeyValue<Key, (Out, Out2)>>> where
Out2: ExchangeData,
OperatorChain2: Operator<KeyValue<Key, Out2>> + 'static,

Given two streams with timestamps join them according to an interval centered around the timestamp of the left side.

This means that an element on the left side with timestamp T will be joined to all the elements on the right with timestamp Q such that T - lower_bound <= Q <= T + upper_bound. Only items with the same key can be joined together.

Note: this operator will split the current block.

Example

TODO: example

impl<Key: DataKey, Out: Data, OperatorChain> KeyedStream<Key, Out, OperatorChain> where
OperatorChain: Operator<KeyValue<Key, Out>> + 'static,

pub fn fold<NewOut: Data, F>(
self,
init: NewOut,
f: F
) -> KeyedStream<Key, NewOut, impl Operator<KeyValue<Key, NewOut>>> where
F: Fn(&mut NewOut, Out) + Send + Clone + 'static,

Perform the folding operation separately for each key.

Note that there is a difference between stream.group_by(keyer).fold(...) and stream.group_by_fold(keyer, ...). The first performs the network shuffle of every item in the stream, and later performs the folding (i.e. nearly all the elements will be sent to the network). The latter avoids sending the items by performing first a local reduction on each host, and then send only the locally folded results (i.e. one message per replica, per key); then the global step is performed aggregating the results.

The resulting stream will still be keyed and will contain only a single message per key (the final result).

Note that the output type may be different from the input type. Consider using KeyedStream::reduce if the output type is the same as the input type.

Note: this operator will retain all the messages of the stream and emit the values only when the stream ends. Therefore this is not properly streaming.

Note: this operator will split the current block.

Example

let s = env.stream(IteratorSource::new((0..5))).group_by(|&n| n % 2);
let res = s
    .fold(0, |acc, value| *acc += value)
    .collect_vec();

env.execute();

let mut res = res.get().unwrap();
res.sort_unstable();
assert_eq!(res, vec![(0, 0 + 2 + 4), (1, 1 + 3)]);

impl<Key: DataKey, Out: Data, OperatorChain> KeyedStream<Key, Out, OperatorChain> where
OperatorChain: Operator<KeyValue<Key, Out>> + 'static,

pub fn reduce<F>(
self,
f: F
) -> KeyedStream<Key, Out, impl Operator<KeyValue<Key, Out>>> where
F: Fn(&mut Out, Out) + Send + Clone + 'static,

Perform the reduction operation separately for each key.

Note that there is a difference between stream.group_by(keyer).reduce(...) and stream.group_by_reduce(keyer, ...). The first performs the network shuffle of every item in the stream, and later performs the reduction (i.e. nearly all the elements will be sent to the network). The latter avoids sending the items by performing first a local reduction on each host, and then send only the locally reduced results (i.e. one message per replica, per key); then the global step is performed aggregating the results.

The resulting stream will still be keyed and will contain only a single message per key (the final result).

Note that the output type must be the same as the input type, if you need a different type consider using KeyedStream::fold.

Note: this operator will retain all the messages of the stream and emit the values only when the stream ends. Therefore this is not properly streaming.

Note: this operator will split the current block.

Example

let s = env.stream(IteratorSource::new((0..5))).group_by(|&n| n % 2);
let res = s
    .reduce(|acc, value| *acc += value)
    .collect_vec();

env.execute();

let mut res = res.get().unwrap();
res.sort_unstable();
assert_eq!(res, vec![(0, 0 + 2 + 4), (1, 1 + 3)]);

impl<Key: DataKey, Out: Data, OperatorChain> KeyedStream<Key, Out, OperatorChain> where
OperatorChain: Operator<KeyValue<Key, Out>> + 'static,

pub fn map<NewOut: Data, F>(
self,
f: F
) -> KeyedStream<Key, NewOut, impl Operator<KeyValue<Key, NewOut>>> where
F: Fn(KeyValue<&Key, Out>) -> NewOut + Send + Clone + 'static,

Map the elements of the stream into new elements.

Note: this is very similar to Iteartor::map.

Example

let s = env.stream(IteratorSource::new((0..5))).group_by(|&n| n % 2);
let res = s.map(|(_key, n)| 10 * n).collect_vec();

env.execute();

let mut res = res.get().unwrap();
res.sort_unstable();
assert_eq!(res, vec![(0, 0), (0, 20), (0, 40), (1, 10), (1, 30)]);

impl<Key: DataKey, Out: Data, OperatorChain> KeyedStream<Key, Out, OperatorChain> where
OperatorChain: Operator<KeyValue<Key, Out>> + 'static,

pub fn rich_filter_map<NewOut: Data, F>(
self,
f: F
) -> KeyedStream<Key, NewOut, impl Operator<KeyValue<Key, NewOut>>> where
F: FnMut(KeyValue<&Key, Out>) -> Option<NewOut> + Send + Clone + 'static,

Remove from the stream all the elements for which the provided function returns None and keep the elements that returned Some(_). The mapping function can be stateful.

This is exactly like Stream::rich_filter_map, but the function is cloned for each key. This means that each key will have a unique mapping function (and therefore a unique state).

impl<Key: DataKey, Out: Data, OperatorChain> KeyedStream<Key, Out, OperatorChain> where
OperatorChain: Operator<KeyValue<Key, Out>> + 'static,

pub fn rich_flat_map<NewOut: Data, MapOut: Data, F>(
self,
f: F
) -> KeyedStream<Key, NewOut, impl Operator<KeyValue<Key, NewOut>>> where
MapOut: IntoIterator<Item = NewOut>,
<MapOut as IntoIterator>::IntoIter: Clone + Send + 'static,
F: FnMut(KeyValue<&Key, Out>) -> MapOut + Clone + Send + 'static,

Apply a mapping operation to each element of the stream, the resulting stream will be the flattened values of the result of the mapping. The mapping function can be stateful.

This is exactly like Stream::rich_flat_map, but the function is cloned for each key. This means that each key will have a unique mapping function (and therefore a unique state).

impl<Key: DataKey, Out: Data, OperatorChain> KeyedStream<Key, Out, OperatorChain> where
OperatorChain: Operator<KeyValue<Key, Out>> + 'static,

pub fn rich_map<NewOut: Data, F>(
self,
f: F
) -> KeyedStream<Key, NewOut, impl Operator<KeyValue<Key, NewOut>>> where
F: FnMut(KeyValue<&Key, Out>) -> NewOut + Clone + Send + 'static,

Map the elements of the stream into new elements. The mapping function can be stateful.

This is exactly like Stream::rich_map, but the function is cloned for each key. This means that each key will have a unique mapping function (and therefore a unique state).

impl<Key: ExchangeDataKey, Out: ExchangeData, OperatorChain> KeyedStream<Key, Out, OperatorChain> where
OperatorChain: Operator<KeyValue<Key, Out>> + 'static,

pub fn collect_vec(self) -> StreamOutput<Vec<(Key, Out)>>

Close the stream and store all the resulting items into a Vec on a single host.

If the stream is distributed among multiple replicas, a bottleneck is placed where all the replicas sends the items to.

Note: the collected items are the pairs (key, value).

Note: the order of items and keys is unspecified.

Note: this operator will split the current block.

Example

let s = env.stream(IteratorSource::new((0..3))).group_by(|&n| n % 2);
let res = s.collect_vec();

env.execute();

let mut res = res.get().unwrap();
res.sort_unstable(); // the output order is nondeterministic
assert_eq!(res, vec![(0, 0), (0, 2), (1, 1)]);

impl<Key: DataKey, Out: Data, OperatorChain> KeyedStream<Key, Out, OperatorChain> where
OperatorChain: Operator<KeyValue<Key, Out>> + 'static,

pub fn for_each<F>(self, f: F) where
F: Fn(Key, Out) + Send + Clone + 'static,

Apply the given function to all the elements of the stream, consuming the stream.

Example

let s = env.stream(IteratorSource::new((0..5))).group_by(|&n| n % 2);
s.for_each(|key, n| println!("Item: {} has key {}", n, key));

env.execute();

impl<Key: DataKey, Out: Data, OperatorChain> KeyedStream<Key, Out, OperatorChain> where
OperatorChain: Operator<KeyValue<Key, Out>> + 'static,

pub fn unkey(
self
) -> Stream<KeyValue<Key, Out>, impl Operator<KeyValue<Key, Out>>>

Make this KeyedStream a normal Stream of key-value pairs.

Example

let stream = env.stream(IteratorSource::new((0..4))).group_by(|&n| n % 2);
let res = stream.unkey().collect_vec();

env.execute();

let mut res = res.get().unwrap();
res.sort_unstable(); // the output order is nondeterministic
assert_eq!(res, vec![(0, 0), (0, 2), (1, 1), (1, 3)]);

impl<Key: DataKey, Out: Data, OperatorChain> KeyedStream<Key, Out, OperatorChain> where
OperatorChain: Operator<KeyValue<Key, Out>> + 'static,

pub fn drop_key(self) -> Stream<Out, impl Operator<Out>>

Forget about the key of this KeyedStream and return a Stream containing just the values.

Example

let stream = env.stream(IteratorSource::new((0..4))).group_by(|&n| n % 2);
let res = stream.drop_key().collect_vec();

env.execute();

let mut res = res.get().unwrap();
res.sort_unstable(); // the output order is nondeterministic
assert_eq!(res, (0..4).collect::<Vec<_>>());

impl<Key: DataKey, Out: Data, OperatorChain> KeyedStream<Key, Out, OperatorChain> where
OperatorChain: Operator<KeyValue<Key, Out>> + 'static,

pub fn window<WinOut: Data, WinDescr: WindowDescription<Key, WinOut>>(
self,
descr: WinDescr
) -> KeyedWindowedStream<Key, Out, impl Operator<KeyValue<Key, Out>>, WinOut, WinDescr>

Apply a window to the stream.

Returns a KeyedWindowedStream, with windows created following the behavior specified by the passed WindowDescription.

Example

let s = env.stream(IteratorSource::new((0..5)));
let res = s
    .group_by(|&n| n % 2)
    .window(CountWindow::sliding(3, 2))
    .sum()
    .collect_vec();

env.execute();

let mut res = res.get().unwrap();
res.sort_unstable();
assert_eq!(res, vec![(0, 4), (0, 0 + 2 + 4), (1, 1 + 3)]);