Struct rstream::Stream

pub struct Stream<Out: Data, OperatorChain> where
    OperatorChain: Operator<Out>,  { /* fields omitted */ }

A Stream represents a chain of operators that work on a flow of data. The type of the elements that is leaving the stream is Out.

Internally a stream is composed by a chain of blocks, each of which can be seen as a simpler stream with input and output types.

A block is internally composed of a chain of operators, nested like the Iterator from std. The type of the chain inside the block is OperatorChain and it’s required as type argument of the stream. This type only represents the chain inside the last block of the stream, not all the blocks inside of it.

Implementations

impl<Out: Data, OperatorChain> Stream<Out, OperatorChain> where
OperatorChain: Operator<Out> + 'static,

pub fn add_timestamps<TimestampGen, WatermarkGen>(
self,
timestamp_gen: TimestampGen,
watermark_gen: WatermarkGen
) -> Stream<Out, impl Operator<Out>> where
TimestampGen: FnMut(&Out) -> Timestamp + Clone + Send + 'static,
WatermarkGen: FnMut(&Out, &Timestamp) -> Option<Timestamp> + Clone + Send + 'static,

Given a stream without timestamps nor watermarks, tag each item with a timestamp and insert watermarks.

The two functions given to this operator are the following:

timestamp_gen returns the timestamp assigned to the provided element of the stream

watermark_gen returns an optional watermark to add after the provided element

Note that the two functions must follow the watermark semantics. TODO: link to watermark semantics

Example

In this example the stream contains the integers from 0 to 9, each will be tagged with a timestamp with the value of the item as milliseconds, and after each even number a watermark will be inserted.

use rstream::operator::Timestamp;

let s = env.stream(IteratorSource::new((0..10)));
s.add_timestamps(
    |&n| Timestamp::from_millis(n),
    |&n, &ts| if n % 2 == 0 { Some(ts) } else { None }
);

impl<Out: ExchangeData, OperatorChain> Stream<Out, OperatorChain> where
OperatorChain: Operator<Out> + 'static,

pub fn group_by_min_element<Key: ExchangeDataKey, Value: Ord, Keyer, GetValue>(
self,
keyer: Keyer,
get_value: GetValue
) -> KeyedStream<Key, Out, impl Operator<KeyValue<Key, Out>>> where
Keyer: KeyerFn<Key, Out> + Fn(&Out) -> Key,
GetValue: KeyerFn<Value, Out> + Fn(&Out) -> Value,

Find, for each partition of the stream, the item with the smallest value.

The stream is partitioned using the keyer function and the value to compare is obtained with get_value.

This operation is associative, therefore the computation is done in parallel before sending all the elements to the network.

Note: the comparison is done using the value returned by get_value, but the resulting items have the same type as the input.

Note: this operator will split the current block.

Example

let s = env.stream(IteratorSource::new((0..5)));
let res = s
    .group_by_min_element(|&n| n % 2, |&n| n)
    .collect_vec();

env.execute();

let mut res = res.get().unwrap();
res.sort_unstable();
assert_eq!(res, vec![(0, 0), (1, 1)]);

impl<Out: ExchangeData, OperatorChain> Stream<Out, OperatorChain> where
OperatorChain: Operator<Out> + 'static,

pub fn group_by_max_element<Key: ExchangeDataKey, Value: Ord, Keyer, GetValue>(
self,
keyer: Keyer,
get_value: GetValue
) -> KeyedStream<Key, Out, impl Operator<KeyValue<Key, Out>>> where
Keyer: KeyerFn<Key, Out> + Fn(&Out) -> Key,
GetValue: KeyerFn<Value, Out> + Fn(&Out) -> Value,

Find, for each partition of the stream, the item with the largest value.

The stream is partitioned using the keyer function and the value to compare is obtained with get_value.

This operation is associative, therefore the computation is done in parallel before sending all the elements to the network.

Note: the comparison is done using the value returned by get_value, but the resulting items have the same type as the input.

Note: this operator will split the current block.

Example

let s = env.stream(IteratorSource::new((0..5)));
let res = s
    .group_by_max_element(|&n| n % 2, |&n| n)
    .collect_vec();

env.execute();

let mut res = res.get().unwrap();
res.sort_unstable();
assert_eq!(res, vec![(0, 4), (1, 3)]);

impl<Out: ExchangeData, OperatorChain> Stream<Out, OperatorChain> where
OperatorChain: Operator<Out> + 'static,

pub fn group_by_sum<Key: ExchangeDataKey, Value, Keyer, GetValue>(
self,
keyer: Keyer,
get_value: GetValue
) -> KeyedStream<Key, Value, impl Operator<KeyValue<Key, Value>>> where
Keyer: KeyerFn<Key, Out> + Fn(&Out) -> Key,
GetValue: KeyerFn<Value, Out> + Fn(&Out) -> Value,
Value: ExchangeData + AddAssign,

Find, for each partition of the stream, the sum of the values of the items.

The stream is partitioned using the keyer function and the value to sum is obtained with get_value.

This operation is associative, therefore the computation is done in parallel before sending all the elements to the network.

Note: this is similar to the SQL: SELECT SUM(value) ... GROUP BY key

Note: the type of the result does not have to be a number, any type that implements AddAssign is accepted.

Note: this operator will split the current block.

Example

let s = env.stream(IteratorSource::new((0..5)));
let res = s
    .group_by_sum(|&n| n % 2, |&n| n)
    .collect_vec();

env.execute();

let mut res = res.get().unwrap();
res.sort_unstable();
assert_eq!(res, vec![(0, 0 + 2 + 4), (1, 1 + 3)]);

impl<Out: ExchangeData, OperatorChain> Stream<Out, OperatorChain> where
OperatorChain: Operator<Out> + 'static,

pub fn group_by_avg<Key: ExchangeDataKey, Value, Keyer, GetValue>(
self,
keyer: Keyer,
get_value: GetValue
) -> KeyedStream<Key, Value, impl Operator<KeyValue<Key, Value>>> where
Keyer: KeyerFn<Key, Out> + Fn(&Out) -> Key,
GetValue: KeyerFn<Value, Out> + Fn(&Out) -> Value,
Value: ExchangeData + AddAssign + Div<f64, Output = Value>,

Find, for each partition of the stream, the average of the values of the items.

The stream is partitioned using the keyer function and the value to average is obtained with get_value.

This operation is associative, therefore the computation is done in parallel before sending all the elements to the network.

Note: this is similar to the SQL: SELECT AVG(value) ... GROUP BY key

Note: the type of the result does not have to be a number, any type that implements AddAssign and can be divided by f64 is accepted.

Note: this operator will split the current block.

Example

let s = env.stream(IteratorSource::new((0..5)));
let res = s
    .group_by_avg(|&n| n % 2, |&n| n as f64)
    .collect_vec();

env.execute();

let mut res = res.get().unwrap();
res.sort_by_key(|(k, _)| *k);
assert_eq!(res, vec![(0, (0.0 + 2.0 + 4.0) / 3.0), (1, (1.0 + 3.0) / 2.0)]);

impl<Out: ExchangeData, OperatorChain> Stream<Out, OperatorChain> where
OperatorChain: Operator<Out> + 'static,

pub fn group_by_count<Key: ExchangeDataKey, Keyer>(
self,
keyer: Keyer
) -> KeyedStream<Key, usize, impl Operator<KeyValue<Key, usize>>> where
Keyer: KeyerFn<Key, Out> + Fn(&Out) -> Key,

Count, for each partition of the stream, the number of items.

The stream is partitioned using the keyer function.

This operation is associative, therefore the computation is done in parallel before sending all the elements to the network.

Note: this is similar to the SQL: SELECT COUNT(*) ... GROUP BY key

Note: this operator will split the current block.

Example

let s = env.stream(IteratorSource::new((0..5)));
let res = s
    .group_by_count(|&n| n % 2)
    .collect_vec();

env.execute();

let mut res = res.get().unwrap();
res.sort_by_key(|(k, _)| *k);
assert_eq!(res, vec![(0, 3), (1, 2)]);

impl<Out: Data, OperatorChain> Stream<Out, OperatorChain> where
OperatorChain: Operator<Out> + 'static,

pub fn batch_mode(self, batch_mode: BatchMode) -> Self

Change the batch mode for this stream.

This change will be propagated to all the operators following, even of the next blocks, until it’s changed again.

Example

use rstream::BatchMode;

let s = env.stream(IteratorSource::new((0..10)));
s.batch_mode(BatchMode::fixed(1024));

impl<Out: ExchangeData, OperatorChain> Stream<Out, OperatorChain> where
OperatorChain: Operator<Out> + 'static,

pub fn broadcast(self) -> Stream<Out, impl Operator<Out>>

Duplicate each element of the stream and forward it to all the replicas of the next block.

Note: this will duplicate the elements of the stream, this is potentially a very expensive operation.

Note: this operator will split the current block.

Example

let s = env.stream(IteratorSource::new((0..10)));
s.broadcast();

impl<Out: ExchangeData, OperatorChain> Stream<Out, OperatorChain> where
OperatorChain: Operator<Out> + 'static,

pub fn concat<OperatorChain2>(
self,
oth: Stream<Out, OperatorChain2>
) -> Stream<Out, impl Operator<Out>> where
OperatorChain2: Operator<Out> + 'static,

Concatenate the items of this stream with the items of another stream with the same type.

Note: the order of the resulting items is not specified.

Note: this operator will split the current block.

Example

let s1 = env.stream(IteratorSource::new((0..10)));
let s2 = env.stream(IteratorSource::new((10..20)));
let res = s1.concat(s2).collect_vec();

env.execute();

let mut res = res.get().unwrap();
res.sort_unstable();
assert_eq!(res, (0..20).collect::<Vec<_>>());

impl<Out: Data, OperatorChain> Stream<Out, OperatorChain> where
OperatorChain: Operator<Out> + 'static,

pub fn filter<Predicate>(
self,
predicate: Predicate
) -> Stream<Out, impl Operator<Out>> where
Predicate: Fn(&Out) -> bool + Clone + Send + 'static,

Remove from the stream all the elements for which the provided predicate returns false.

Note: this is very similar to Iteartor::filter

Example

let s = env.stream(IteratorSource::new((0..10)));
let res = s.filter(|&n| n % 2 == 0).collect_vec();

env.execute();

assert_eq!(res.get().unwrap(), vec![0, 2, 4, 6, 8])

impl<Out: Data, OperatorChain> Stream<Out, OperatorChain> where
OperatorChain: Operator<Out> + 'static,

pub fn filter_map<NewOut: Data, F>(
self,
f: F
) -> Stream<NewOut, impl Operator<NewOut>> where
F: Fn(Out) -> Option<NewOut> + Send + Clone + 'static,

Remove from the stream all the elements for which the provided function returns None and keep the elements that returned Some(_).

Note: this is very similar to Iteartor::filter_map

Example

let s = env.stream(IteratorSource::new((0..10)));
let res = s.filter_map(|n| if n % 2 == 0 { Some(n * 3) } else { None }).collect_vec();

env.execute();

assert_eq!(res.get().unwrap(), vec![0, 6, 12, 18, 24])

impl<Out: Data, OperatorChain, NewOut: Data> Stream<Out, OperatorChain> where
Out: IntoIterator<Item = NewOut>,
<Out as IntoIterator>::IntoIter: Clone + Send + 'static,
OperatorChain: Operator<Out> + 'static,

pub fn flatten(self) -> Stream<NewOut, impl Operator<NewOut>>

Transform this stream of containers into a stream of all the contained values.

Note: this is very similar to Iteartor::flatten

Example

let s = env.stream(IteratorSource::new(vec![
    vec![1, 2, 3],
    vec![],
    vec![4, 5],
].into_iter()));
let res = s.flatten().collect_vec();

env.execute();

assert_eq!(res.get().unwrap(), vec![1, 2, 3, 4, 5]);

impl<Out: Data, OperatorChain> Stream<Out, OperatorChain> where
OperatorChain: Operator<Out> + 'static,

pub fn flat_map<MapOut: Data, NewOut: Data, F>(
self,
f: F
) -> Stream<NewOut, impl Operator<NewOut>> where
MapOut: IntoIterator<Item = NewOut>,
<MapOut as IntoIterator>::IntoIter: Clone + Send + 'static,
F: Fn(Out) -> MapOut + Send + Clone + 'static,

Apply a mapping operation to each element of the stream, the resulting stream will be the flattened values of the result of the mapping.

Note: this is very similar to Iteartor::flat_map

Example

let s = env.stream(IteratorSource::new((0..3)));
let res = s.flat_map(|n| vec![n, n]).collect_vec();

env.execute();

assert_eq!(res.get().unwrap(), vec![0, 0, 1, 1, 2, 2]);

impl<Out: Data, OperatorChain> Stream<Out, OperatorChain> where
OperatorChain: Operator<Out> + 'static,

pub fn fold<NewOut: Data, F>(
self,
init: NewOut,
f: F
) -> Stream<NewOut, impl Operator<NewOut>> where
Out: ExchangeData,
F: Fn(&mut NewOut, Out) + Send + Clone + 'static,

Fold the stream into a stream that emits a single value.

The folding operator consists in adding to the current accumulation value (initially the value provided as init) the value of the current item in the stream.

The folding function is provided with a mutable reference to the current accumulator and the owned item of the stream. The function should modify the accumulator without returning anything.

Note that the output type may be different from the input type. Consider using Stream::reduce if the output type is the same as the input type.

Note: this operator will retain all the messages of the stream and emit the values only when the stream ends. Therefore this is not properly streaming.

Note: this operator is not parallelized, it creates a bottleneck where all the stream elements are sent to and the folding is done using a single thread.

Note: this is very similar to Iteartor::fold.

Note: this operator will split the current block.

Example

let s = env.stream(IteratorSource::new((0..5)));
let res = s.fold(0, |acc, value| *acc += value).collect_vec();

env.execute();

assert_eq!(res.get().unwrap(), vec![0 + 1 + 2 + 3 + 4]);

pub fn fold_assoc<NewOut: ExchangeData, Local, Global>(
self,
init: NewOut,
local: Local,
global: Global
) -> Stream<NewOut, impl Operator<NewOut>> where
Local: Fn(&mut NewOut, Out) + Send + Clone + 'static,
Global: Fn(&mut NewOut, NewOut) + Send + Clone + 'static,

Fold the stream into a stream that emits a single value.

The folding operator consists in adding to the current accumulation value (initially the value provided as init) the value of the current item in the stream.

This method is very similary to Stream::fold, but performs the folding distributely. To do so the folding function must be associative, in particular the folding process is performed in 2 steps:

local: the local function is used to fold the elements present in each replica of the stream independently. All those replicas will start with the same init value.

global: all the partial results (the elements produced by the local step) have to be aggregated into a single result. This is done using the global folding function.

Note that the output type may be different from the input type, therefore requireing different function for the aggregation. Consider using Stream::reduce_assoc if the output type is the same as the input type.

Note: this operator will retain all the messages of the stream and emit the values only when the stream ends. Therefore this is not properly streaming.

Note: this operator will split the current block.

Example

let s = env.stream(IteratorSource::new((0..5)));
let res = s.fold_assoc(0, |acc, value| *acc += value, |acc, value| *acc += value).collect_vec();

env.execute();

assert_eq!(res.get().unwrap(), vec![0 + 1 + 2 + 3 + 4]);

impl<Out: ExchangeData, OperatorChain> Stream<Out, OperatorChain> where
OperatorChain: Operator<Out> + 'static,

pub fn group_by<Key: DataKey, Keyer>(
self,
keyer: Keyer
) -> KeyedStream<Key, Out, impl Operator<KeyValue<Key, Out>>> where
Keyer: Fn(&Out) -> Key + Send + Clone + 'static,

Given a stream, make a KeyedStream partitioning the values according to a key generated by the keyer function provided.

The returned KeyedStream is partitioned by key, and all the operators added to it will be evaluated after the network shuffle. Therefore all the items are sent to the network (if their destination is not the local host). In many cases this behaviour can be avoided by using the associative variant of the operators (e.g. Stream::group_by_reduce, Stream::group_by_sum, …).

Note: the keys are not sent to the network, they are built on the sending side, and rebuilt on the receiving side.

Note: this operator will split the current block.

Example

let s = env.stream(IteratorSource::new((0..5)));
let keyed = s.group_by(|&n| n % 2); // partition even and odd elements

impl<Out: ExchangeData, OperatorChain> Stream<Out, OperatorChain> where
OperatorChain: Operator<Out> + 'static,

pub fn interval_join<Out2, OperatorChain2>(
self,
right: Stream<Out2, OperatorChain2>,
lower_bound: Duration,
upper_bound: Duration
) -> Stream<(Out, Out2), impl Operator<(Out, Out2)>> where
Out2: ExchangeData,
OperatorChain2: Operator<Out2> + 'static,

Given two streams with timestamps join them according to an interval centered around the timestamp of the left side.

This means that an element on the left side with timestamp T will be joined to all the elements on the right with timestamp Q such that T - lower_bound <= Q <= T + upper_bound.

Note: this operator is not parallelized, all the elements are sent to a single node to perform the join.

Note: this operator will split the current block.

Example

TODO: example

impl<Out: ExchangeData, OperatorChain> Stream<Out, OperatorChain> where
OperatorChain: Operator<Out> + 'static,

pub fn iterate<Body, DeltaUpdate: ExchangeData + Default, State: ExchangeData + Sync, OperatorChain2>(
self,
num_iterations: usize,
initial_state: State,
body: Body,
local_fold: impl Fn(&mut DeltaUpdate, Out) + Send + Clone + 'static,
global_fold: impl Fn(&mut State, DeltaUpdate) + Send + Clone + 'static,
loop_condition: impl Fn(&mut State) -> bool + Send + Clone + 'static
) -> (Stream<State, impl Operator<State>>, Stream<Out, impl Operator<Out>>) where
Body: FnOnce(Stream<Out, Iterate<Out, State, OperatorChain>>, IterationStateHandle<State>) -> Stream<Out, OperatorChain2>,
OperatorChain2: Operator<Out> + 'static,

Construct an iterative dataflow where the input stream is fed inside a cycle. What comes out of the loop will be fed back at the next iteration.

This iteration is stateful, this means that all the replicas have a read-only access to the iteration state. The initial value of the state is given as parameter. When an iteration ends all the elements are reduced locally at each replica producing a DeltaUpdate. Those delta updates are later reduced on a single node that, using the global_fold function will compute the state for the next iteration. This state is also used in loop_condition to check whether the next iteration should start or not. loop_condition is also allowed to mutate the state.

The initial value of DeltaUpdate is initialized with Default::default().

The content of the loop has a new scope: it’s defined by the body function that takes as parameter the stream of data coming inside the iteration and a reference to the state. This function should return the stream of the data that exits from the loop (that will be fed back).

This construct produces two stream:

the first is a stream with a single item: the final state of the iteration

the second if the set of elements that exited the loop during the last iteration (i.e. the ones that should have been fed back in the next iteration).

Note: due to an internal limitation, it’s not currently possible to add an iteration operator when the stream has limited parallelism. This means, for example, that after a non-parallel source you have to add a shuffle.

Note: this operator will split the current block.

Example

let s = env.stream(IteratorSource::new(0..3)).shuffle();
let (state, items) = s.iterate(
    3, // at most 3 iterations
    0, // the initial state is zero
    |s, state| s.map(|n| n + 10),
    |delta: &mut i32, n| *delta += n,
    |state, delta| *state += delta,
    |_state| true,
);
let state = state.collect_vec();
let items = items.collect_vec();
env.execute();

assert_eq!(state.get().unwrap(), vec![10 + 11 + 12 + 20 + 21 + 22 + 30 + 31 + 32]);
assert_eq!(items.get().unwrap(), vec![30, 31, 32]);

impl<Out: Data, OperatorChain> Stream<Out, OperatorChain> where
OperatorChain: Operator<Out> + 'static,

pub fn replay<Body, DeltaUpdate: ExchangeData + Default, State: ExchangeData + Sync, OperatorChain2>(
self,
num_iterations: usize,
initial_state: State,
body: Body,
local_fold: impl Fn(&mut DeltaUpdate, Out) + Send + Clone + 'static,
global_fold: impl Fn(&mut State, DeltaUpdate) + Send + Clone + 'static,
loop_condition: impl Fn(&mut State) -> bool + Send + Clone + 'static
) -> Stream<State, impl Operator<State>> where
Body: FnOnce(Stream<Out, Replay<Out, State, OperatorChain>>, IterationStateHandle<State>) -> Stream<Out, OperatorChain2>,
OperatorChain2: Operator<Out> + 'static,

Construct an iterative dataflow where the input stream is repeatedly fed inside a cycle, i.e. what comes into the cycle is replayed at every iteration.

This iteration is stateful, this means that all the replicas have a read-only access to the iteration state. The initial value of the state is given as parameter. When an iteration ends all the elements are reduced locally at each replica producing a DeltaUpdate. Those delta updates are later reduced on a single node that, using the global_fold function will compute the state for the next iteration. This state is also used in loop_condition to check whether the next iteration should start or not. loop_condition is also allowed to mutate the state.

The initial value of DeltaUpdate is initialized with Default::default().

The content of the loop has a new scope: it’s defined by the body function that takes as parameter the stream of data coming inside the iteration and a reference to the state. This function should return the stream of the data that exits from the loop (that will be fed back).

This construct produces a single stream with a single element: the final state of the iteration.

Note: due to an internal limitation, it’s not currently possible to add an iteration operator when the stream has limited parallelism. This means, for example, that after a non-parallel source you have to add a shuffle.

Note: this operator will split the current block.

Example

let s = env.stream(IteratorSource::new(0..3)).shuffle();
let state = s.replay(
    3, // at most 3 iterations
    0, // the initial state is zero
    |s, state| s.map(|n| n + 10),
    |delta: &mut i32, n| *delta += n,
    |state, delta| *state += delta,
    |_state| true,
);
let state = state.collect_vec();
env.execute();

assert_eq!(state.get().unwrap(), vec![3 * (10 + 11 + 12)]);

impl<Out: ExchangeData, OperatorChain> Stream<Out, OperatorChain> where
OperatorChain: Operator<Out> + 'static,

pub fn join<Out2: ExchangeData, OperatorChain2, Key, Keyer1, Keyer2>(
self,
rhs: Stream<Out2, OperatorChain2>,
keyer1: Keyer1,
keyer2: Keyer2
) -> KeyedStream<Key, InnerJoinTuple<Out, Out2>, impl Operator<KeyValue<Key, InnerJoinTuple<Out, Out2>>>> where
Key: DataKey,
OperatorChain2: Operator<Out2> + 'static,
Keyer1: Fn(&Out) -> Key + KeyerFn<Key, Out>,
Keyer2: Fn(&Out2) -> Key + KeyerFn<Key, Out2>,

Given two stream, create a stream with all the pairs (left item from the left stream, right item from the right), such that the key obtained with keyer1 on an item from the left is equal to the key obtained with keyer2 on an item from the right.

This is an inner join, very similar to SELECT a, b FROM a JOIN b ON keyer1(a) = keyer2(b).

This is a shortcut for: self.join_with(...).ship_hash().local_hash().inner().

Note: this operator will split the current block.

Example

let s1 = env.stream(IteratorSource::new(0..5u8));
let s2 = env.stream(IteratorSource::new(0..5i32));
let res = s1.join(s2, |n| (n % 5) as i32, |n| n % 2).drop_key().collect_vec();

env.execute();

let mut res = res.get().unwrap();
res.sort_unstable();
assert_eq!(res, vec![(0, 0), (0, 2), (0, 4), (1, 1), (1, 3)]);

pub fn left_join<Out2: ExchangeData, OperatorChain2, Key, Keyer1, Keyer2>(
self,
rhs: Stream<Out2, OperatorChain2>,
keyer1: Keyer1,
keyer2: Keyer2
) -> KeyedStream<Key, LeftJoinTuple<Out, Out2>, impl Operator<KeyValue<Key, LeftJoinTuple<Out, Out2>>>> where
Key: DataKey,
OperatorChain2: Operator<Out2> + 'static,
Keyer1: Fn(&Out) -> Key + KeyerFn<Key, Out>,
Keyer2: Fn(&Out2) -> Key + KeyerFn<Key, Out2>,

Given two stream, create a stream with all the pairs (left item from the left stream, right item from the right), such that the key obtained with keyer1 on an item from the left is equal to the key obtained with keyer2 on an item from the right.

This is a left join, meaning that if an item from the left does not find and element from the right with which make a pair, an extra pair (left, None) is generated. If you want to have a right join, you just need to switch the two sides and use a left join.

This is very similar to SELECT a, b FROM a LEFT JOIN b ON keyer1(a) = keyer2(b).

This is a shortcut for: self.join_with(...).ship_hash().local_hash().left().

Note: this operator will split the current block.

Example

let s1 = env.stream(IteratorSource::new(0..5u8));
let s2 = env.stream(IteratorSource::new(0..5i32));
let res = s1.left_join(s2, |n| (n % 5) as i32, |n| n % 2).drop_key().collect_vec();

env.execute();

let mut res = res.get().unwrap();
res.sort_unstable();
assert_eq!(res, vec![(0, Some(0)), (0, Some(2)), (0, Some(4)), (1, Some(1)), (1, Some(3)), (2, None), (3, None), (4, None)]);

pub fn outer_join<Out2: ExchangeData, OperatorChain2, Key, Keyer1, Keyer2>(
self,
rhs: Stream<Out2, OperatorChain2>,
keyer1: Keyer1,
keyer2: Keyer2
) -> KeyedStream<Key, OuterJoinTuple<Out, Out2>, impl Operator<KeyValue<Key, OuterJoinTuple<Out, Out2>>>> where
Key: DataKey,
OperatorChain2: Operator<Out2> + 'static,
Keyer1: Fn(&Out) -> Key + KeyerFn<Key, Out>,
Keyer2: Fn(&Out2) -> Key + KeyerFn<Key, Out2>,

Given two stream, create a stream with all the pairs (left item from the left stream, right item from the right), such that the key obtained with keyer1 on an item from the left is equal to the key obtained with keyer2 on an item from the right.

This is a full-outer join, meaning that if an item from the left does not find and element from the right with which make a pair, an extra pair (left, None) is generated. Similarly if an element from the right does not appear in any pair, a new one is generated with (None, right).

This is very similar to SELECT a, b FROM a FULL OUTER JOIN b ON keyer1(a) = keyer2(b).

This is a shortcut for: self.join_with(...).ship_hash().local_hash().outer().

Note: this operator will split the current block.

Example

let s1 = env.stream(IteratorSource::new(0..5u8));
let s2 = env.stream(IteratorSource::new(0..5i32));
let res = s1.outer_join(s2, |n| (n % 5) as i32, |n| n % 2).drop_key().collect_vec();

env.execute();

let mut res = res.get().unwrap();
res.sort_unstable();
assert_eq!(res, vec![(Some(0), Some(0)), (Some(0), Some(2)), (Some(0), Some(4)), (Some(1), Some(1)), (Some(1), Some(3)), (Some(2), None), (Some(3), None), (Some(4), None)]);

pub fn join_with<Out2: ExchangeData, OperatorChain2, Key, Keyer1, Keyer2>(
self,
rhs: Stream<Out2, OperatorChain2>,
keyer1: Keyer1,
keyer2: Keyer2
) -> JoinStream<Key, Out, Out2, OperatorChain, OperatorChain2, Keyer1, Keyer2> where
OperatorChain2: Operator<Out2>,
Keyer1: Fn(&Out) -> Key + KeyerFn<Key, Out>,
Keyer2: Fn(&Out2) -> Key + KeyerFn<Key, Out2>,

Given two streams, start building a join operator.

The returned type allows you to customize the behaviour of the join. You can select which ship strategy and which local strategy to use.

Ship strategies

hash: the hash of the key is used to select where to send the elements

broadcast right: the left stream is left locally, while the right stream is broadcasted

Local strategies

hash: build an hashmap to match the tuples

sort and merge: collect all the tuples, sort them by key and merge them

The first strategy to select is the ship strategy. After choosing that you have to select the local strategy.

Example

let s1 = env.stream(IteratorSource::new(0..5u8));
let s2 = env.stream(IteratorSource::new(0..5i32));
let j = s1.join_with(s2, |n| (n % 5) as i32, |n| n % 2).ship_hash();

let s1 = env.stream(IteratorSource::new(0..5u8));
let s2 = env.stream(IteratorSource::new(0..5i32));
let j = s1.join_with(s2, |n| (n % 5) as i32, |n| n % 2).ship_broadcast_right();

impl<Out: Data, OperatorChain> Stream<Out, OperatorChain> where
OperatorChain: Operator<Out> + 'static,

pub fn key_by<Key: DataKey, Keyer>(
self,
keyer: Keyer
) -> KeyedStream<Key, Out, impl Operator<KeyValue<Key, Out>>> where
Keyer: Fn(&Out) -> Key + Send + Clone + 'static,

Construct a KeyedStream from a Stream without shuffling the data.

Note: this violates the semantics of KeyedStream, without sending all the values with the same key to the same replica some of the following operators may misbehave. You probably need to use Stream::group_by instead.

Example

let s = env.stream(IteratorSource::new((0..5)));
let res = s.key_by(|&n| n % 2).collect_vec();

env.execute();

let mut res = res.get().unwrap();
res.sort_unstable();
assert_eq!(res, vec![(0, 0), (0, 2), (0, 4), (1, 1), (1, 3)]);

impl<Out: Data, OperatorChain> Stream<Out, OperatorChain> where
OperatorChain: Operator<Out> + 'static,

pub fn group_by_fold<Key: ExchangeDataKey, NewOut: ExchangeData, Keyer, Local, Global>(
self,
keyer: Keyer,
init: NewOut,
local: Local,
global: Global
) -> KeyedStream<Key, NewOut, impl Operator<KeyValue<Key, NewOut>>> where
Keyer: Fn(&Out) -> Key + Send + Clone + 'static,
Local: Fn(&mut NewOut, Out) + Send + Clone + 'static,
Global: Fn(&mut NewOut, NewOut) + Send + Clone + 'static,

Perform the folding operation separately for each key.

This is equivalent of partitioning the stream using the keyer function, and then applying Stream::fold_assoc to each partition separately.

Note however that there is a difference between stream.group_by(keyer).fold(...) and stream.group_by_fold(keyer, ...). The first performs the network shuffle of every item in the stream, and later performs the folding (i.e. nearly all the elements will be sent to the network). The latter avoids sending the items by performing first a local reduction on each host, and then send only the locally folded results (i.e. one message per replica, per key); then the global step is performed aggregating the results.

The resulting stream will still be keyed and will contain only a single message per key (the final result).

Note that the output type may be different from the input type, therefore requireing different function for the aggregation. Consider using Stream::group_by_reduce if the output type is the same as the input type.

Note: this operator will retain all the messages of the stream and emit the values only when the stream ends. Therefore this is not properly streaming.

Note: this operator will split the current block.

Example

let s = env.stream(IteratorSource::new((0..5)));
let res = s
    .group_by_fold(|&n| n % 2, 0, |acc, value| *acc += value, |acc, value| *acc += value)
    .collect_vec();

env.execute();

let mut res = res.get().unwrap();
res.sort_unstable();
assert_eq!(res, vec![(0, 0 + 2 + 4), (1, 1 + 3)]);

impl<Out: ExchangeData, OperatorChain> Stream<Out, OperatorChain> where
OperatorChain: Operator<Out> + 'static,

pub fn group_by_reduce<Key: ExchangeDataKey, Keyer, F>(
self,
keyer: Keyer,
f: F
) -> KeyedStream<Key, Out, impl Operator<KeyValue<Key, Out>>> where
Keyer: Fn(&Out) -> Key + Send + Clone + 'static,
F: Fn(&mut Out, Out) + Send + Clone + 'static,

Perform the reduction operation separately for each key.

This is equivalent of partitioning the stream using the keyer function, and then applying Stream::reduce_assoc to each partition separately.

Note however that there is a difference between stream.group_by(keyer).reduce(...) and stream.group_by_reduce(keyer, ...). The first performs the network shuffle of every item in the stream, and later performs the reduction (i.e. nearly all the elements will be sent to the network). The latter avoids sending the items by performing first a local reduction on each host, and then send only the locally reduced results (i.e. one message per replica, per key); then the global step is performed aggregating the results.

The resulting stream will still be keyed and will contain only a single message per key (the final result).

Note that the output type must be the same as the input type, if you need a different type consider using Stream::group_by_fold.

Note: this operator will retain all the messages of the stream and emit the values only when the stream ends. Therefore this is not properly streaming.

Note: this operator will split the current block.

Example

let s = env.stream(IteratorSource::new((0..5)));
let res = s
    .group_by_reduce(|&n| n % 2, |acc, value| *acc += value)
    .collect_vec();

env.execute();

let mut res = res.get().unwrap();
res.sort_unstable();
assert_eq!(res, vec![(0, 0 + 2 + 4), (1, 1 + 3)]);

impl<Out: Data, OperatorChain> Stream<Out, OperatorChain> where
OperatorChain: Operator<Out> + 'static,

pub fn map<NewOut: Data, F>(self, f: F) -> Stream<NewOut, impl Operator<NewOut>> where
F: Fn(Out) -> NewOut + Send + Clone + 'static,

Map the elements of the stream into new elements.

Note: this is very similar to Iteartor::map.

Example

let s = env.stream(IteratorSource::new((0..5)));
let res = s.map(|n| n * 10).collect_vec();

env.execute();

assert_eq!(res.get().unwrap(), vec![0, 10, 20, 30, 40]);

impl<Out: ExchangeData, OperatorChain> Stream<Out, OperatorChain> where
OperatorChain: Operator<Out> + 'static,

pub fn max_parallelism(
self,
max_parallelism: usize
) -> Stream<Out, impl Operator<Out>>

Change the maximum parallelism of the following operators.

Note: this operator is pretty advanced, some operators may need to be fully replicated and will fail otherwise.

impl<Out: ExchangeData, OperatorChain> Stream<Out, OperatorChain> where
OperatorChain: Operator<Out> + 'static,

pub fn reduce<F>(self, f: F) -> Stream<Out, impl Operator<Out>> where
F: Fn(&mut Out, Out) + Send + Clone + 'static,

Reduce the stream into a stream that emits a single value.

The reducing operator consists in adding to the current accumulation value the value of the current item in the stream.

The reducing function is provided with a mutable reference to the current accumulator and the owned item of the stream. The function should modify the accumulator without returning anything.

Note that the output type must be the same as the input type, if you need a different type consider using Stream::fold.

Note: this operator will retain all the messages of the stream and emit the values only when the stream ends. Therefore this is not properly streaming.

Note: this operator is not parallelized, it creates a bottleneck where all the stream elements are sent to and the folding is done using a single thread.

Note: this is very similar to Iteartor::reduce.

Note: this operator will split the current block.

Example

let s = env.stream(IteratorSource::new((0..5)));
let res = s.reduce(|acc, value| *acc += value).collect_vec();

env.execute();

assert_eq!(res.get().unwrap(), vec![0 + 1 + 2 + 3 + 4]);

pub fn reduce_assoc<F>(self, f: F) -> Stream<Out, impl Operator<Out>> where
F: Fn(&mut Out, Out) + Send + Clone + 'static,

Reduce the stream into a stream that emits a single value.

The reducing operator consists in adding to the current accumulation value the value of the current item in the stream.

This method is very similary to Stream::reduce, but performs the reduction distributely. To do so the reducing function must be associative, in particular the reducing process is performed in 2 steps:

local: the reducing function is used to reduce the elements present in each replica of the stream independently.

global: all the partial results (the elements produced by the local step) have to be aggregated into a single result.

Note that the output type must be the same as the input type, if you need a different type consider using Stream::fold_assoc.

Note: this operator will retain all the messages of the stream and emit the values only when the stream ends. Therefore this is not properly streaming.

Note: this operator will split the current block.

Example

let s = env.stream(IteratorSource::new((0..5)));
let res = s.reduce_assoc(|acc, value| *acc += value).collect_vec();

env.execute();

assert_eq!(res.get().unwrap(), vec![0 + 1 + 2 + 3 + 4]);

impl<Out: Data, OperatorChain> Stream<Out, OperatorChain> where
OperatorChain: Operator<Out> + 'static,

pub fn rich_filter_map<NewOut: Data, F>(
self,
f: F
) -> Stream<NewOut, impl Operator<NewOut>> where
F: FnMut(Out) -> Option<NewOut> + Send + Clone + 'static,

Remove from the stream all the elements for which the provided function returns None and keep the elements that returned Some(_). The mapping function can be stateful.

This is equivalent to Stream::filter_map but with a stateful function.

Since the mapping function can be stateful, it is a FnMut. This allows expressing simple algorithms with very few lines of code (see examples).

The mapping function is cloned inside each replica, and they will not share state between each other. If you want that only a single replica handles all the items you may want to change the parallelism of this operator with Stream::max_parallelism.

Examples

This will emit only the positive prefix-sums.

let s = env.stream(IteratorSource::new(std::array::IntoIter::new([1, 2, -5, 3, 1])));
let res = s.rich_filter_map({
    let mut sum = 0;
    move |x| {
        sum += x;
        if sum >= 0 {
            Some(sum)
        } else {
            None
        }
    }
}).collect_vec();

env.execute();

assert_eq!(res.get().unwrap(), vec![1, 1 + 2, /* 1 + 2 - 5, */ 1 + 2 - 5 + 3, 1 + 2 - 5 + 3 + 1]);

impl<Out: Data, OperatorChain> Stream<Out, OperatorChain> where
OperatorChain: Operator<Out> + 'static,

pub fn rich_flat_map<MapOut: Data, NewOut: Data, F>(
self,
f: F
) -> Stream<NewOut, impl Operator<NewOut>> where
MapOut: IntoIterator<Item = NewOut>,
<MapOut as IntoIterator>::IntoIter: Clone + Send + 'static,
F: FnMut(Out) -> MapOut + Send + Clone + 'static,

Apply a mapping operation to each element of the stream, the resulting stream will be the flattened values of the result of the mapping. The mapping function can be stateful.

This is equivalent to Stream::flat_map but with a stateful function.

Since the mapping function can be stateful, it is a FnMut. This allows expressing simple algorithms with very few lines of code (see examples).

The mapping function is cloned inside each replica, and they will not share state between each other. If you want that only a single replica handles all the items you may want to change the parallelism of this operator with Stream::max_parallelism.

Examples

This will emit only the positive prefix-sums.

let s = env.stream(IteratorSource::new((0..=3)));
let res = s.rich_flat_map({
    let mut elements = Vec::new();
    move |y| {
        let new_pairs = elements
            .iter()
            .map(|&x: &u32| (x, y))
            .collect::<Vec<_>>();
        elements.push(y);
        new_pairs
    }
}).collect_vec();

env.execute();

assert_eq!(res.get().unwrap(), vec![(0, 1), (0, 2), (1, 2), (0, 3), (1, 3), (2, 3)]);

impl<Out: Data, OperatorChain> Stream<Out, OperatorChain> where
OperatorChain: Operator<Out> + 'static,

pub fn rich_map<NewOut: Data, F>(
self,
f: F
) -> Stream<NewOut, impl Operator<NewOut>> where
F: FnMut(Out) -> NewOut + Send + Clone + 'static,

Map the elements of the stream into new elements. The mapping function can be stateful.

This is equivalent to Stream::map but with a stateful function.

Since the mapping function can be stateful, it is a FnMut. This allows expressing simple algorithms with very few lines of code (see examples).

The mapping function is cloned inside each replica, and they will not share state between each other. If you want that only a single replica handles all the items you may want to change the parallelism of this operator with Stream::max_parallelism.

Examples

This is a simple implementation of the prefix-sum using a single replica (i.e. each element is mapped to the sum of all the elements up to that point). Note that this won’t work if there are more replicas.

let s = env.stream(IteratorSource::new((1..=5)));
let res = s.rich_map({
    let mut sum = 0;
    move |x| {
        sum += x;
        sum
    }
}).collect_vec();

env.execute();

assert_eq!(res.get().unwrap(), vec![1, 1 + 2, 1 + 2 + 3, 1 + 2 + 3 + 4, 1 + 2 + 3 + 4 + 5]);

This will enumerate all the elements that reach a replica. This is basically equivalent to the enumerate function in Python.

let s = env.stream(IteratorSource::new((1..=5)));
let res = s.rich_map({
    let mut id = 0;
    move |x| {
        id += 1;
        (id - 1, x)
    }
}).collect_vec();

env.execute();

assert_eq!(res.get().unwrap(), vec![(0, 1), (1, 2), (2, 3), (3, 4), (4, 5)]);

impl<Out: ExchangeData, OperatorChain> Stream<Out, OperatorChain> where
OperatorChain: Operator<Out> + 'static,

pub fn shuffle(self) -> Stream<Out, impl Operator<Out>>

Perform a network shuffle sending the messages to a random replica.

This can be useful if for some reason the load is very unbalanced (e.g. after a very unbalanced Stream::group_by).

Note: this operator will split the current block.

Example

let s = env.stream(IteratorSource::new((0..5)));
let res = s.shuffle();

impl<Out: ExchangeData, OperatorChain> Stream<Out, OperatorChain> where
OperatorChain: Operator<Out> + 'static,

pub fn collect_vec(self) -> StreamOutput<Vec<Out>>

Close the stream and store all the resulting items into a Vec on a single host.

If the stream is distributed among multiple replicas, a bottleneck is placed where all the replicas sends the items to.

Note: the order of items and keys is unspecified.

Note: this operator will split the current block.

Example

let s = env.stream(IteratorSource::new((0..10)));
let res = s.collect_vec();

env.execute();

assert_eq!(res.get().unwrap(), (0..10).collect::<Vec<_>>());

impl<Out: Data, OperatorChain> Stream<Out, OperatorChain> where
OperatorChain: Operator<Out> + 'static,

pub fn for_each<F>(self, f: F) where
F: FnMut(Out) + Send + Clone + 'static,

Apply the given function to all the elements of the stream, consuming the stream.

Example

let s = env.stream(IteratorSource::new((0..5)));
s.for_each(|n| println!("Item: {}", n));

env.execute();

impl<Out: ExchangeData, OperatorChain> Stream<Out, OperatorChain> where
OperatorChain: Operator<Out> + 'static,

pub fn split(self, splits: usize) -> Vec<Stream<Out, impl Operator<Out>>>

Split the stream into splits streams, each with all the elements of the first one.

This will effectively duplicate every item in the stream into the newly created streams.

Note: this operator will split the current block.

Example

let s = env.stream(IteratorSource::new((0..5)));
let mut splits = s.split(3);
let a = splits.pop().unwrap();
let b = splits.pop().unwrap();
let c = splits.pop().unwrap();

impl<Out: ExchangeData, OperatorChain> Stream<Out, OperatorChain> where
OperatorChain: Operator<Out> + 'static,

pub fn window_all<WinOut: Data, WinDescr: WindowDescription<(), WinOut>>(
self,
descr: WinDescr
) -> WindowedStream<Out, impl Operator<KeyValue<(), Out>>, WinOut, WinDescr>

Apply a window to the stream.

Returns a WindowedStream, with windows created following the behavior specified by the passed WindowDescription.

Note: this operator cannot be parallelized, so all the stream elements are sent to a single node where the creation and aggregation of the windows are done.

Example

let s = env.stream(IteratorSource::new((0..5)));
let res = s
    .window_all(CountWindow::tumbling(2))
    .sum()
    .collect_vec();

env.execute();

let mut res = res.get().unwrap();
assert_eq!(res, vec![0 + 1, 2 + 3, 4]);

impl<Out1: ExchangeData, OperatorChain1> Stream<Out1, OperatorChain1> where
OperatorChain1: Operator<Out1> + 'static,

pub fn zip<Out2: ExchangeData, OperatorChain2>(
self,
oth: Stream<Out2, OperatorChain2>
) -> Stream<(Out1, Out2), impl Operator<(Out1, Out2)>> where
OperatorChain2: Operator<Out2> + 'static,

Given two Streams, zip their elements together: the resulting stream will be a stream of pairs, each of which is an element from both streams respectively.

Note: all the elements after the end of one of the streams are discarded (i.e. the resulting stream will have a number of elements that is the minimum between the lengths of the two input streams).

Note: this operator will split the current block.

Example

let s1 = env.stream(IteratorSource::new(vec!['A', 'B', 'C', 'D'].into_iter()));
let s2 = env.stream(IteratorSource::new(vec![1, 2, 3].into_iter()));
let res = s1.zip(s2).collect_vec();

env.execute();

assert_eq!(res.get().unwrap(), vec![('A', 1), ('B', 2), ('C', 3)]);

impl<Out: Data, OperatorChain> Stream<Out, OperatorChain> where
OperatorChain: Operator<Out> + 'static,

pub fn add_operator<NewOut: Data, Op, GetOp>(
self,
get_operator: GetOp
) -> Stream<NewOut, Op> where
Op: Operator<NewOut> + 'static,
GetOp: FnOnce(OperatorChain) -> Op,

Add a new operator to the current chain inside the stream. This consumes the stream and returns a new one with the operator added.

get_operator is a function that is given the previous chain of operators and should return the new chain of operators. The new chain cannot be simply passed as argument since it is required to do a partial move of the InnerBlock structure.

Note: this is an advanced function that manipulates the block structure. Probably it is not what you are looking for.