Struct rstream::WindowedStream

pub struct WindowedStream<Out: Data, OperatorChain, WinOut: Data, WinDescr> where
    OperatorChain: Operator<KeyValue<(), Out>>,
    WinDescr: WindowDescription<(), WinOut>,  { /* fields omitted */ }

A WindowedStream is a data stream where elements are divided in multiple groups called windows. Internally, a WindowedStream is just a KeyedWindowedStream where each element is assigned to the same key ().

These are the windows supported out-of-the-box:

EventTimeWindow

EventTimeWindow::sliding

EventTimeWindow::tumbling

EventTimeWindow::session

ProcessingTimeWindow

ProcessingTimeWindow::sliding

ProcessingTimeWindow::tumbling

ProcessingTimeWindow::session

CountWindow

CountWindow::sliding

CountWindow::tumbling

To apply a window to a Stream, see Stream::window_all.

Implementations

impl<Out: Data, WindowDescr, OperatorChain> WindowedStream<Out, OperatorChain, Out, WindowDescr> where
WindowDescr: WindowDescription<(), Out> + Clone + 'static,
OperatorChain: Operator<KeyValue<(), Out>> + 'static,

pub fn first(self) -> Stream<Out, impl Operator<Out>>

Returns the first element of each window.

Example

let s = env.stream(IteratorSource::new((0..5)));
let res = s
    .window_all(CountWindow::sliding(3, 2))
    .first()
    .collect_vec();

env.execute();

let mut res = res.get().unwrap();
assert_eq!(res, vec![0, 2, 4]);

impl<Out: Data, WindowDescr, OperatorChain> WindowedStream<Out, OperatorChain, Out, WindowDescr> where
WindowDescr: WindowDescription<(), Out> + Clone + 'static,
OperatorChain: Operator<KeyValue<(), Out>> + 'static,

pub fn fold<NewOut: Data, F>(
self,
init: NewOut,
fold: F
) -> Stream<NewOut, impl Operator<NewOut>> where
F: Fn(&mut NewOut, &Out) + Clone + Send + 'static,

Folds the elements of each window into an accumulator value

fold() takes two arguments: the initial value of the accumulator and a closure used to accumulate the elements of each window.

The closure is called once for each element of each window with two arguments: a mutable reference to the accumulator and the element of the window. The closure should modify the accumulator, without returning anything.

Example

let s = env.stream(IteratorSource::new((0..5)));
let res = s
    .window_all(CountWindow::sliding(3, 2))
    .fold(1, |acc, &n| *acc *= n)
    .collect_vec();

env.execute();

let mut res = res.get().unwrap();
assert_eq!(res, vec![(0 * 1 * 2), (2 * 3 * 4), (4)]);

impl<Out: ExchangeData, Out2: ExchangeData, WindowDescr, OperatorChain> WindowedStream<Out, OperatorChain, ConcatElement<Out, Out2>, WindowDescr> where
WindowDescr: WindowDescription<(), ConcatElement<Out, Out2>> + Clone + 'static,
OperatorChain: Operator<KeyValue<(), Out>> + 'static,

pub fn join<OperatorChain2>(
self,
right: Stream<Out2, OperatorChain2>
) -> Stream<(Out, Out2), impl Operator<(Out, Out2)>> where
OperatorChain2: Operator<Out2> + 'static,

Joins the elements coming from two streams into pairs.

Joins on windows are equivalent to:

Merging the two streams into a single one.

Grouping the elements in windows, following the rules of the passed window description.

For each window, generating a tuple for each pair of elements coming from the two streams.

This join behaves as an inner join.

Also see Stream::interval_join.

Example

let s1 = env
    .stream(IteratorSource::new((0..4)))
    .add_timestamps(
        |&n| Timestamp::from_millis(n),
        |&n, &ts| Some(ts)
    );
let s2 = env
    .stream(IteratorSource::new((4..8)))
    .add_timestamps(
        |&n| Timestamp::from_millis(n - 4),
        |&n, &ts| Some(ts)
    );
let res = s1
    .window_all(EventTimeWindow::tumbling(Duration::from_millis(2)))
    .join(s2)
    .collect_vec();

env.execute();

let mut res = res.get().unwrap();
res.sort_unstable();
assert_eq!(res, vec![(0, 4), (0, 5), (1, 4), (1, 5), (2, 6), (2, 7), (3, 6), (3, 7)]);

impl<Out: Data, WindowDescr, OperatorChain> WindowedStream<Out, OperatorChain, Out, WindowDescr> where
WindowDescr: WindowDescription<(), Out> + Clone + 'static,
OperatorChain: Operator<KeyValue<(), Out>> + 'static,

pub fn map<NewOut: Data, F>(
self,
map_func: F
) -> Stream<NewOut, impl Operator<NewOut>> where
F: Fn(&mut dyn ExactSizeIterator<Item = &Out>) -> NewOut + Clone + Send + 'static,

Takes a closure used to transform each window into a new element. The closure is called once for each window, passing an iterator over the elements of the window as argument.

Returns a Stream containing all the values returned by the closure.

Example

let s = env.stream(IteratorSource::new((0..5)));
let res = s
    .window_all(CountWindow::sliding(3, 2))
    .map(|window| *window.last().unwrap())
    .collect_vec();

env.execute();

let mut res = res.get().unwrap();
assert_eq!(res, vec![2, 4, 4]);

impl<Out: Data + Ord, WindowDescr, OperatorChain> WindowedStream<Out, OperatorChain, Out, WindowDescr> where
WindowDescr: WindowDescription<(), Out> + Clone + 'static,
OperatorChain: Operator<KeyValue<(), Out>> + 'static,

pub fn max(self) -> Stream<Out, impl Operator<Out>>

Returns the maximum element of each window.

Example

let s = env.stream(IteratorSource::new((0..5)));
let res = s
    .window_all(CountWindow::sliding(3, 2))
    .max()
    .collect_vec();

env.execute();

let mut res = res.get().unwrap();
assert_eq!(res, vec![0.max(1).max(2), 2.max(3).max(4), 4]);

impl<Out: Data + Ord, WindowDescr, OperatorChain> WindowedStream<Out, OperatorChain, Out, WindowDescr> where
WindowDescr: WindowDescription<(), Out> + Clone + 'static,
OperatorChain: Operator<KeyValue<(), Out>> + 'static,

pub fn min(self) -> Stream<Out, impl Operator<Out>>

Returns the minimum element of each window.

Example

let s = env.stream(IteratorSource::new((0..5)));
let res = s
    .window_all(CountWindow::sliding(3, 2))
    .min()
    .collect_vec();

env.execute();

let mut res = res.get().unwrap();
assert_eq!(res, vec![0.min(1).min(2), 2.min(3).min(4), 4]);

impl<Out: Data + Ord, WindowDescr, OperatorChain> WindowedStream<Out, OperatorChain, Out, WindowDescr> where
WindowDescr: WindowDescription<(), Out> + Clone + 'static,
OperatorChain: Operator<KeyValue<(), Out>> + 'static,
for<'a> Out: Sum<&'a Out>,

pub fn sum(self) -> Stream<Out, impl Operator<Out>>

Sums all the elements of each window.

Example

let s = env.stream(IteratorSource::new((0..5)));
let res = s
    .window_all(CountWindow::sliding(3, 2))
    .sum()
    .collect_vec();

env.execute();

let mut res = res.get().unwrap();
assert_eq!(res, vec![0 + 1 + 2, 2 + 3 + 4, 4]);