Module rstream::structure

This mod contains the types and function that model the structure of the blocks and operators for debugging purposes only.

Structs

BlockStructure
The structural information about a block.

Connection
A connection registered by an operator.

DataType
Wrapper type that contains a string representing the type.

OperatorReceiver
A receiver registered by an operator.

OperatorStructure
The structural information about an operator.

Enums

ConnectionStrategy
The strategy used for sending the data in a channel.

OperatorKind
The kind of operator: either Operator, Source or Sink.