Module rstream::operator::join

Some utility structures for building the join operators.

The actual operators are Stream::join, Stream::left_join, Stream::outer_join and Stream::join_with.

Structs

JoinStream
Intermediate stream type for building the join between two streams.

JoinStreamLocalHash
This is an intermediate type for building a join operator.

JoinStreamLocalSortMerge
This is an intermediate type for building a join operator.

JoinStreamShipBroadcastRight
This is an intermediate type for building a join operator.

JoinStreamShipHash
This is an intermediate type for building a join operator.

ShipBroadcastRight
Marker type for remembering that broadcast_right is the selected ship strategy.

ShipHash
Marker type for remembering that hash is the selected ship strategy.

Traits

ShipStrategy
Marker trait for the ship strategy marker types.

Type Definitions

InnerJoinTuple
Type alias for a pair of joined items in an inner join.

LeftJoinTuple
Type alias for a pair of joined items in a left join.

OuterJoinTuple
Type alias for a pair of joined items in an outer join.