Trait rstream::operator::Operator

pub trait Operator<Out: Data>: Clone {
    fn setup(&mut self, metadata: ExecutionMetadata);

    fn next(&mut self) -> StreamElement<Out>;

    fn to_string(&self) -> String;

    fn structure(&self) -> BlockStructure;
}

An operator represents a unit of computation. It's always included inside a chain of operators, inside a block.

Each operator implements the Operator<Out> trait, it produced a stream of Out elements.

An Operator must be Clone since it is part of a single chain when it's built, but it has to be cloned to spawn the replicas of the block.

Required methods

fn setup(&mut self, metadata: ExecutionMetadata)

Setup the operator chain. This is called before any call to next and it's used to initialize the operator. When it's called the operator has already been cloned and it will never be cloned again. Therefore it's safe to store replica-specific metadata inside of it.

It's important that each operator (except the start of a chain) calls .setup() recursively on the previous operators.

fn next(&mut self) -> StreamElement<Out>

Take a value from the previous operator, process it and return it.

fn to_string(&self) -> String

A string representation of the operator and its predecessors.

fn structure(&self) -> BlockStructure

A more refined representation of the operator and its predecessors.

Loading content...